requirements.txt
python-dotenv
requests
openai
bing-image-urls
argparse
beautifulsoup4
google
